Refactor ResearchLoop into specialized manager classes
- Extracted PlanGenerator, ExecutionManager, ReflectionManager, and ReportGenerator from ResearchLoop.
- Moved split_text_into_chunks to deep_research_project/core/utils.py.
- Updated ResearchLoop to delegate responsibilities to these new managers.
- Added new module files: planning.py, execution.py, reflection.py, reporting.py, utils.py.
- Updated tests to reflect the new structure.

This refactor addresses the Single Responsibility Principle violation in ResearchLoop, making the code more modular, maintainable, and testable.
